Moving CFFFont into cff folder
